requirements.txt
absl-py==1.1.0
adal==1.2.7
affine==2.3.0
aiohttp==3.8.3
aiosignal==1.3.1
ale-py==0.7.3
altgraph==0.17.2
anyio==3.1.0
appdirs==1.4.4
appnope==0.1.2
apptools==5.2.0
APScheduler==3.6.3
argon2-cffi==20.1.0
arrow==1.2.2
asgiref==3.4.1
aspose-cells==21.12.1
astropy==5.2.2
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
atari-py==0.2.9
atomicwrites==1.4.0
attrs==21.2.0
audioread==2.1.9
azure-common==1.1.28
azure-core==1.22.1
azure-identity==1.12.0
azure-mgmt-core==1.3.2
azure-mgmt-subscription==3.1.1
azure-storage-blob==12.11.0
Babel==2.9.1
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.2.2
beautifulsoup4==4.9.3
bleach==3.3.0
bokeh==2.4.2
branca==0.6.0
cachetools==4.2.2
captum==0.5.0
catboost==0.25.1
certifi==2021.5.30
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
click-plugins==1.1.1
cligj==0.7.1
cloudpickle==2.1.0
colorama==0.4.4
colour==0.1.5
configobj==5.0.8
contourpy==1.0.6
covid==2.0.7
cryptography==3.4.8
cssutils==2.6.0
cycler==0.10.0
dash==2.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataframe-image==0.1.11
DateTime==4.4
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.0.8
django-allauth==0.45.0
django-cors-headers==3.7.0
django-filter==21.1
django-formset-js-improved==0.5.0.2
django-jquery-js==3.1.1
django-jsonforms==1.1.2
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
dnspython==2.2.1
docker-pycreds==0.4.0
docx==0.2.4
dtreeviz==2.2.1
entrypoints==0.3
envisage==6.1.1
et-xmlfile==1.1.0
ffmpeg==1.4
filelock==3.8.0
Flask==2.1.2
flatbuffers==1.12
folium==0.13.0
fonttools==4.38.0
frozenlist==1.3.3
future==0.18.2
fuzzywuzzy==0.18.0
gast==0.4.0
geoip2==4.6.0
gitdb==4.0.9
GitPython==3.1.27
glob2==0.7
google==3.0.0
google-api-core==1.31.3
google-api-python-client==1.12.8
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.53.0
graphviz==0.16
grpcio==1.32.0
gspread==5.7.1
gunicorn==20.0.4
gym==0.21.0
h11==0.13.0
h5py==2.10.0
html2image==2.0.3
html5lib==1.1
httpcore==0.16.2
httplib2==0.20.1
httpx==0.23.1
idna==2.8
imageio==2.9.0
importlib-metadata==4.8.1
importlib-resources==5.4.0
iniconfig==1.1.1
ipykernel==5.5.5
ipython==7.23.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.0.0
joblib==1.0.1
JPype1==1.3.0
json5==0.9.5
jsons==1.6.2
jsonschema==4.0.1
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-server==1.8.0
jupyterlab==3.0.15
jupyterlab-pygments==0.1.2
jupyterlab-server==2.5.1
jupyterlab-widgets==1.0.0
kaleido==0.2.1
keras==2.9.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kneed==0.8.2
KqlmagicCustom==0.1.114.post22
libclang==14.0.1
librosa==0.8.0
linecache2==1.0.0
llvmlite==0.36.0
lxml==4.8.0
Markdown==3.3.4
MarkupSafe==2.1.1
marshmallow==3.14.0
marshmallow-jsonschema==0.12.0
matplotlib==3.6.2
matplotlib-inline==0.1.2
maxminddb==2.2.0
mayavi==4.8.1
mistune==0.8.4
mongoengine==0.22.1
mongomock==3.21.0
mpld3==0.5.7
msal==1.20.0
msal-extensions==1.0.0
msrest==0.6.21
msrestazure==0.6.4
msticpy==2.2.0
multidict==6.0.2
munch==2.5.0
musicalbeeps==0.2.9
names==0.3.0
nbclassic==0.2.8
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==3.0rc1
nmmn==1.2.1
notebook==6.3.0
numba==0.53.1
numpy==1.21.5
oauth2client==4.1.3
oauthlib==3.1.1
opencv-contrib-python==4.5.1.48
opencv-python==4.4.0.46
openpyxl==3.0.9
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.0
pandas==1.3.5
pandocfilters==1.4.3
paramiko==2.12.0
parso==0.8.2
pathlib==1.0.1
pathtools==0.1.2
pefile==2021.9.3
pexpect==4.8.0
pickle-mixin==1.0.2
pickleshare==0.7.5
Pillow==9.0.1
playsound==1.3.0
plotly==5.13.1
pluggy==0.13.1
pooch==1.3.0
portalocker==2.6.0
prettytable==3.5.0
prometheus-client==0.10.1
promise==2.3
prompt-toolkit==3.0.18
protobuf==3.17.3
psutil==5.8.0
psycopg2-binary==2.8.5
ptyprocess==0.7.0
py==1.10.0
py2exe==0.11.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.12.0
pydantic==1.8.2
pydotplus==2.0.2
PyDrive==1.3.1
pydub==0.25.1
pyerfa==2.0.0.3
pyface==7.4.4
pyglet==1.5.21
Pygments==2.9.0
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.7
PyJWT==2.1.0
pymongo==3.12.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyperclip==1.8.2
pyproj==3.0.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
pyrsistent==0.18.0
pyserial==3.5
PySocks==1.7.1
pytest==6.2.3
python-dateutil==2.7.5
python-docx==0.8.11
python-gitlab==2.5.0
python-pptx==0.6.21
python-telegram-bot==13.7
python-time==0.3.0
python3-openid==3.2.0
pytz==2019.3
PyWavelets==1.1.1
pywin32==300
pywin32-ctypes==0.2.0
pywinpty==1.1.1
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.1.0
QtPy==1.9.0
Random-Word==1.0.11
requests==2.27.1
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
resampy==0.2.2
retrying==1.3.3
rfc3986==1.5.0
rsa==4.7.2
rstr==2.2.6
scapy==2.4.5
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.7.3
selenium==4.1.5
Send2Trash==1.5.0
sentinels==1.0.0
sentry-sdk==1.5.11
setproctitle==1.2.3
shap==0.41.0
Shapely==1.7.1
shortuuid==1.0.9
simpleaudio==1.0.4
six==1.16.0
sklearn==0.0
slicer==0.0.7
smmap==5.0.0
sniffio==1.2.0
snuggs==1.4.7
sortedcontainers==2.4.0
sounddevice==0.4.1
SoundFile==0.10.3.post1
soupsieve==2.3.1
sqlparse==0.4.2
stable-baselines===2.10.3a0.-WIP-
stable-baselines3==1.5.0
telegram==0.0.1
tenacity==8.1.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-gpu==2.9.1
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.10.0
testpath==0.5.0
threadpoolctl==2.1.0
tifffile==2021.4.8
tldextract==3.4.0
toml==0.10.2
torch==1.8.1
torchvision==0.9.1
tornado==6.1
tqdm==4.63.1
traceback2==1.4.0
traitlets==5.0.5
traits==6.4.1
traitsui==7.4.3
trio==0.20.0
trio-websocket==0.9.2
typing-extensions==3.10.0.2
typish==1.9.3
tzdata==2021.2.post0
tzlocal==3.0
uritemplate==3.0.1
urllib3==1.26.9
utils==1.0.1
vobject==0.9.6.1
vtk==9.2.6
wandb==0.13.6
wcwidth==0.2.5
webdriver-manager==3.5.4
webencodings==0.5.1
websocket-client==1.0.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
wsproto==1.1.0
xgboost==1.4.2
xlrd==2.0.1
XlsxWriter==3.0.3
xmltodict==0.12.0
yarl==1.8.1
zipp==3.6.0
zope.interface==5.4.0
